Build for amd64 and armv6l and arm64

+++ sysstat.go
@@ -28,7 +28,7 @@ func sysstatGet() (sysstat, error) {
 	var sysinfo unix.Sysinfo_t
 	err := unix.Sysinfo(&sysinfo)
 	if err == nil {
-		res.Uptime = sysinfo.Uptime
+		res.Uptime = int64(sysinfo.Uptime)
 	}
 
 	var usage unix.Rusage
